Move the sub-info into the info namespace.
