+++ osu.Game.Rulesets.Karaoke/Edit/Lyrics/Rows/Info/Badge/LanguageInfo.cs
@@ -17,7 +17,7 @@ using osu.Game.Rulesets.Karaoke.Edit.Lyrics.States;
 using osu.Game.Rulesets.Karaoke.Objects;
 using osu.Game.Rulesets.Karaoke.Utils;
 
-namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Rows.Components.SubInfo
+namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Rows.Info.Badge
 {
     public class LanguageInfo : SubInfo, IHasPopover
     {
