+++ osu.Game.Rulesets.Karaoke/Edit/Lyrics/Rows/Info/Badge/TimeTagInfo.cs
@@ -9,7 +9,7 @@ using osu.Game.Graphics;
 using osu.Game.Rulesets.Karaoke.Objects;
 using osu.Game.Rulesets.Karaoke.Utils;
 
-namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Rows.Components.SubInfo
+namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Rows.Info.Badge
 {
     public class TimeTagInfo : SubInfo
     {
